Removed some compiler warnings
git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@108 a27c51de-2d2b-4d1b-a6e8-f7f87814f453

+++ projects/arrayOptimization/ArrayProcessor.C
@@ -28,7 +28,7 @@ extern bool DebugAnnot();
 
 bool ToArrayOnly( const vector<string>& argv )
 {
-  for (int i = 1; i < argv.size(); ++i) {
+  for (size_t i = 1; i < argv.size(); ++i) {
     if ( argv[i] == "-toarrayonly")
         return true;
   }
